+++ CHANGES.rst
@@ -1,6 +1,11 @@
 11.16.17 (unreleased)
 =====================
 
+Roman
+-----
+
+- Dataset to Ref header key matching where "roman" prefix is missing [#910]  
+
 General
 -------
 - exclude build/ and install.log from source control [#907]